requirements.txt
pandas
numpy
requests
psycopg2-binary
selenium
flask
chromedriver-autoinstaller
python-dotenv
gunicorn
schedule

